requirements.txt
# ---------- Core Framework ----------
fastapi==0.128.0
starlette==0.50.0
uvicorn==0.40.0
anyio==4.12.1
h11==0.16.0

# ---------- Database (Async MySQL) ----------
SQLAlchemy==2.0.46
aiomysql==0.3.2
PyMySQL==1.1.2
greenlet==3.3.1

# ---------- Security ----------
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.5.0
rsa==4.9.1
ecdsa==0.19.1
pyasn1==0.6.2
six==1.17.0

# ---------- Environment & Config ----------
python-dotenv==1.2.1

# ---------- Validation / Typing ----------
pydantic==2.12.5
pydantic_core==2.41.5
annotated-types==0.7.0
typing_extensions==4.15.0
typing-inspection==0.4.2

# ---------- Utilities ----------
click==8.3.1
colorama==0.4.6
idna==3.11

# ---------- Docs / Optional ----------
annotated-doc==0.0.4


pydantic[email]